+++ acm_website/src/components/Navbar.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import '../styles/NavBar.css';
-import { useApp } from '../contexts/AppContext';
+import { useApp } from '../hooks/useApp';
 import logo from '../assets/logo.png';
 
 const Navbar: React.FC = () => {
