+++ acm_website/src/pages/AdminPage.tsx
@@ -7,7 +7,7 @@ import Account from '../components/admin/Account';
 import { auth } from '../firebase/config';
 import { signOut } from "firebase/auth";
 import { collection, doc, getFirestore, getDocs, query, where, updateDoc, addDoc, Timestamp, orderBy, arrayUnion } from "firebase/firestore";
-import { useApp } from '../contexts/AppContext';
+import { useApp } from '../hooks/useApp';
 import * as XLSX from 'xlsx';
 import { deleteUser } from '../api';
